added github build workflow

+++ client/src/Contexts/eventContext.tsx
@@ -3,15 +3,15 @@ import eventApi from "../api/eventApi";
 
 type EventsContextType = {
     events: IEvent[];
-    createEvent: (date: Date, name: string) => void;
-    updateEvent: (eventId: string, newDate: Date, newName: string) => void;
+    createEvent: (date: Date, name: string, imageFile: File) => Promise<boolean>;
+    updateEvent: (eventId: string, newDate: Date, newName: string) => Promise<boolean>;
     deleteEvent: (eventId: string) => void;
 };
 
 const EventContext = React.createContext<EventsContextType | undefined>(undefined);
 
 const EventProvider: React.FC<{ children: React.ReactNode }> = ({ children }) => {
-    const [ events, setEvents] = React.useState<IEvent[]>([]);
+    const [ events, setEvents ] = React.useState<IEvent[]>([]);
 
     const getEventFromId = (eventId: String): IEvent => {
         const event: IEvent | undefined = events.find((event) => event.id === eventId);
@@ -25,42 +25,48 @@ const EventProvider: React.FC<{ children: React.ReactNode }> = ({ children }) =>
 
     const fetchEvents = async () => {
         try {
-            const response = await eventApi.fetchEvents();
-            setEvents(response.data);
+            const events: IEvent[] = await eventApi.fetchEvents();
+            setEvents(events);
         } catch (error) {
             console.error("Failed to fetch events:", error);
         }
     };
 
-    const createEvent = async (date: Date, name: string) => {
+    const createEvent = async (date: Date, name: string, imageFile: File): Promise<boolean> => {
         try {
-            await eventApi.createEvent(date, name);
+            await eventApi.createEvent(date, name, imageFile);
             fetchEvents();
+            return true;
         }
         catch (error) {
             console.error("Failed to create event:", error);
         }
+        return false;
     };
 
-    const updateEvent = async (eventId: string, newDate: Date, newName: string) => {
+    const updateEvent = async (eventId: string, newDate: Date, newName: string): Promise<boolean> => {
         const updatedEvent: Partial<IEvent> = { };
         const currentEvent: IEvent = getEventFromId(eventId);
 
         if (newDate !== currentEvent.date) updatedEvent.date = newDate;
         if (newName !== currentEvent.name) updatedEvent.name = newName;
 
-        if (Object.keys(updatedEvent).length === 0) return; // No changes to update
+        if (Object.keys(updatedEvent).length === 0) return false; // No changes to update
 
         try {
             eventApi.updateEvent(eventId, updatedEvent);
             fetchEvents();
+            return true
         }
         catch (error) {
             console.error("Failed to update event:", error);
         }
+        return false;
     };
 
     const deleteEvent = async (eventId: string) => {
+        if (!window.confirm("Are you sure you want to delete this event?")) return;
+        
         try {
             await eventApi.deleteEvent(eventId);
             fetchEvents();
